Placing orders when signal is generated
